set apiDocumentContentTypeMiddleware on prod server
