+++ _config.ts
@@ -7,7 +7,6 @@ import search from "lume/plugins/search.ts";
 import esbuild from "lume/plugins/esbuild.ts";
 import redirects from "lume/plugins/redirects.ts";
 import sitemap from "lume/plugins/sitemap.ts";
-import type { Middleware } from "lume/core/server.ts";
 
 import tailwindConfig from "./tailwind.config.js";
 
@@ -24,7 +23,6 @@ import codeblockTitlePlugin from "./markdown-it/codeblock-title.ts";
 import toc from "https://deno.land/x/lume_markdown_plugins@v0.7.0/toc.ts";
 import title from "https://deno.land/x/lume_markdown_plugins@v0.7.0/title.ts";
 import { CSS as GFM_CSS } from "https://jsr.io/@deno/gfm/0.8.2/style.ts";
-import { parseMediaType } from "https://jsr.io/@std/media-types/1.0.1/parse_media_type.ts";
 import {
   deploy as oramaDeploy,
   generateDocumentsForExamples,
@@ -32,37 +30,7 @@ import {
   generateDocumentsForSymbols,
   OramaDocument,
 } from "./orama.ts";
-
-function isCssOrJs(contentType: string | null): boolean {
-  if (contentType === null) {
-    return false;
-  }
-
-  const [mime] = parseMediaType(contentType);
-  return ["text/css", "text/javascript", "application/javascript"].includes(
-    mime,
-  );
-}
-
-const apiDocumentContentTypeMiddleware: Middleware = async (request, next) => {
-  const response = await next(request);
-
-  // Requests to API docs located under `/api` should all be served as HTML with
-  // a few exceptions like `/api/deno/page.css` or `/api/web/script.js`.
-  // We need to explicitly set the content type here because some file names
-  // confuse the MIME type detection - for example, the file name
-  // `Deno.Kv.prototype.list` is inferred as `text/plain` but it should be
-  // `text/html` for browsers to render it correctly.
-  const requestToApiDoc = new URL(request.url).pathname.startsWith("/api");
-  const contentTypeSetToCssOrJs = isCssOrJs(
-    response.headers.get("content-type"),
-  );
-  if (requestToApiDoc && !contentTypeSetToCssOrJs) {
-    response.headers.set("content-type", "text/html");
-  }
-
-  return response;
-};
+import { apiDocumentContentTypeMiddleware } from "./middleware.ts";
 
 const site = lume({
   location: new URL("https://docs.deno.com"),
@@ -107,6 +75,7 @@ site.copy("deploy/kv/tutorials/images");
 site.copy("runtime/manual/images");
 site.copy("deno.json");
 site.copy("server.ts");
+site.copy("middleware.ts");
 
 site.use(redirects({
   output: "json",
